Generator refactored
